refactor(mmrelay): enforce stricter typing for config and constants
apply Final typing to config keys and related constants

replace mutable tuples with frozenset for environment booleans

update default constants to explicit Final annotations

adjust tests and CI/docs to reflect the new typings and keys

minor logging and formatting tweaks across modules for consistency

+++ tests/constants.py
@@ -61,7 +61,7 @@ TEST_SHORT_DELAY: Final[float] = 0.1
 TEST_HALF_SECOND: Final[float] = 0.5
 TEST_NETWORK_TIMEOUT: Final[float] = 5.0
 TEST_PLUGIN_TIMEOUT: Final[float] = 5.0
-TEST_GIT_TIMEOUT: Final[int] = 120
+TEST_GIT_TIMEOUT: Final[float] = 120.0
 TEST_LONG_TIMEOUT: Final[float] = 10.0
 
 # Test SQL snippets for database testing
